backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syllabus-parser-backend"
version = "0.1.0"
description = "Backend for Syllabus Parser — extracts assignments from syllabi"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pymupdf>=1.24",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "twilio>=9.0",
    "openai>=1.14",
    "supabase>=2.4",
    "pydantic-settings>=2.2",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "black>=24.3",
    "flake8>=7.0",
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
